+++ crates/dav/src/card/proppatch.rs
@@ -49,7 +49,7 @@ pub(crate) trait CardPropPatchRequestHandler: Sync + Send {
 
     fn apply_addressbook_properties(
         &self,
-        access_token: &AccessToken,
+        personal_id: u32,
         address_book: &mut AddressBook,
         is_update: bool,
         properties: Vec<DavPropertyValue>,
@@ -156,11 +156,12 @@ impl CardPropPatchRequestHandler for Server {
             let mut new_book = archive
                 .deserialize::<AddressBook>()
                 .caused_by(trc::location!())?;
+            let personal_id = access_token.personal_id(account_id, Collection::AddressBook);
 
             // Remove properties
             if !request.set_first && !request.remove.is_empty() {
                 remove_addressbook_properties(
-                    access_token,
+                    personal_id,
                     &mut new_book,
                     std::mem::take(&mut request.remove),
                     &mut items,
@@ -169,7 +170,7 @@ impl CardPropPatchRequestHandler for Server {
 
             // Set properties
             is_success = self.apply_addressbook_properties(
-                access_token,
+                personal_id,
                 &mut new_book,
                 true,
                 request.set,
@@ -179,7 +180,7 @@ impl CardPropPatchRequestHandler for Server {
             // Remove properties
             if is_success && !request.remove.is_empty() {
                 remove_addressbook_properties(
-                    access_token,
+                    personal_id,
                     &mut new_book,
                     request.remove,
                     &mut items,
@@ -261,7 +262,7 @@ impl CardPropPatchRequestHandler for Server {
 
     fn apply_addressbook_properties(
         &self,
-        access_token: &AccessToken,
+        personal_id: u32,
         address_book: &mut AddressBook,
         is_update: bool,
         properties: Vec<DavPropertyValue>,
@@ -273,7 +274,7 @@ impl CardPropPatchRequestHandler for Server {
             match (&property.property, property.value) {
                 (DavProperty::WebDav(WebDavProperty::DisplayName), DavValue::String(name)) => {
                     if name.len() <= self.core.groupware.live_property_size {
-                        address_book.preferences_mut(access_token).name = name;
+                        address_book.preferences_mut(personal_id).name = name;
                         items.insert_ok(property.property);
                     } else {
                         items.insert_error_with_description(
@@ -289,7 +290,7 @@ impl CardPropPatchRequestHandler for Server {
                     DavValue::String(name),
                 ) => {
                     if name.len() <= self.core.groupware.live_property_size {
-                        address_book.preferences_mut(access_token).description = Some(name);
+                        address_book.preferences_mut(personal_id).description = Some(name);
                         items.insert_ok(property.property);
                     } else {
                         items.insert_error_with_description(
@@ -456,7 +457,7 @@ fn remove_card_properties(
 }
 
 fn remove_addressbook_properties(
-    access_token: &AccessToken,
+    personal_id: u32,
     book: &mut AddressBook,
     properties: Vec<DavProperty>,
     items: &mut PropStatBuilder,
@@ -464,11 +465,11 @@ fn remove_addressbook_properties(
     for property in properties {
         match &property {
             DavProperty::CardDav(CardDavProperty::AddressbookDescription) => {
-                book.preferences_mut(access_token).description = None;
+                book.preferences_mut(personal_id).description = None;
                 items.insert_with_status(property, StatusCode::NO_CONTENT);
             }
             DavProperty::WebDav(WebDavProperty::DisplayName) => {
-                book.preferences_mut(access_token).name.clear();
+                book.preferences_mut(personal_id).name.clear();
                 items.insert_with_status(property, StatusCode::NO_CONTENT);
             }
             DavProperty::DeadProperty(dead) => {