+++ crates/jmap/src/addressbook/get.rs
@@ -49,6 +49,7 @@ impl AddressBookGet for Server {
             AddressBookProperty::MyRights,
         ]);
         let account_id = request.account_id.document_id();
+        let personal_id = access_token.personal_id(account_id, Collection::AddressBook);
         let cache = self
             .fetch_dav_resources(
                 access_token.account_id(),
@@ -128,14 +129,14 @@ impl AddressBookGet for Server {
                     AddressBookProperty::Name => {
                         result.insert_unchecked(
                             AddressBookProperty::Name,
-                            address_book.preferences(access_token).name.to_string(),
+                            address_book.preferences(personal_id).name.to_string(),
                         );
                     }
                     AddressBookProperty::Description => {
                         result.insert_unchecked(
                             AddressBookProperty::Description,
                             address_book
-                                .preferences(access_token)
+                                .preferences(personal_id)
                                 .description
                                 .as_ref()
                                 .map(|v| v.to_string()),
@@ -144,10 +145,7 @@ impl AddressBookGet for Server {
                     AddressBookProperty::SortOrder => {
                         result.insert_unchecked(
                             AddressBookProperty::SortOrder,
-                            address_book
-                                .preferences(access_token)
-                                .sort_order
-                                .to_native(),
+                            address_book.preferences(personal_id).sort_order.to_native(),
                         );
                     }
                     AddressBookProperty::IsDefault => {
@@ -162,7 +160,7 @@ impl AddressBookGet for Server {
                             address_book
                                 .subscribers
                                 .iter()
-                                .any(|account_id| *account_id == access_token.account_id()),
+                                .any(|subscriber| *subscriber == personal_id),
                         );
                     }
                     AddressBookProperty::ShareWith => {